Update README.md
This is a Profit Analysis Project I undertook during my training with The Incubator Hub. This project focuses on different international breweries brand and the profit generated across different regions and countries within 2017-2019 in other to make evaluative decisions about where to scale and where to concentrate more marketing efforts.